#1124: move the duplicated per-family Variables overrides into the shared base
`GeneratorMoveNextEmitter.Variables` and `AsyncGeneratorMoveNextEmitter.Variables`
carried near-byte-identical block-scope-rename + function-display-class routing:
`RenameToken`, `TryGetFunctionDCField`, `StoreToDCField`, and the nine
`EmitVariable` / `EmitVarDeclaration` / `EmitAssign` / `EmitStoreVariable` /
`EmitConstDeclaration` / `EmitCompoundAssign` / `EmitLogicalAssign` /
`EmitPrefixIncrement` / `EmitPostfixIncrement` overrides.

Move them into the shared `IteratorMoveNextEmitter` base (added in #1123). The
only family-specific input — the per-binding `BlockScopeRenames` map, drawn from
each family's own analysis — becomes an abstract property each subclass provides
(the sync generator returns it directly; the async generator keeps its
`?? NoRenames` fallback). The shared methods reach the display-class field through
the existing `GetFunctionDCField()` hook (still overridden per family, since it
touches the concrete `_builder`) and use the base `IL`/`Ctx`/`Types` accessors
instead of the concrete fields, so the emitted IL is unchanged.

The generator keeps its own `ResolveCaptureSourceName` (async generators do not
support async arrows, so that capture-source pivot is generator-only).

Pure code move — byte-identical IL. Build green; closure/capture/generator
targeted suite 1385/1385.
